create separate logging component and separate portal component
